[REFACTOR] Moved home activity test code to fragment
