+++ app/src/main/java/com/example/android/barebone/ui/home/homefragments/FragmentA.kt
@@ -1,19 +1,30 @@
 package com.example.android.barebone.ui.home.homefragments
 
+import android.content.Intent
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import androidx.databinding.DataBindingUtil
 import androidx.fragment.app.Fragment
+import androidx.lifecycle.Observer
 import androidx.lifecycle.ViewModelProvider
 import androidx.lifecycle.ViewModelProviders
 import com.example.android.barebone.R
+import com.example.android.barebone.databinding.FragmentDemoContentABinding
 import com.example.android.barebone.di.Injectable
+import com.example.android.barebone.ui.featurex.FeatureXActivity
+import com.example.android.barebone.ui.featurey.FeatureYActivity
+import com.example.android.barebone.ui.featurez.FeatureZActivity
 import timber.log.Timber
 import javax.inject.Inject
 
 /**
  * Demo fragment for tab content.
+ *
+ * This fragment is different than [FragmentB] and [FragmentC] in following ways:
+ * - This has it's own view model [FragmentAViewModel] which is provided via injected factory.
+ * - This loads a custom layout to test other activities.
  *
  * TODO: Move the fragment to it's own feature package.
  */
@@ -21,6 +32,7 @@ class FragmentA : Fragment(), Injectable {
     @Inject
     lateinit var viewModelFactory: ViewModelProvider.Factory
 
+    lateinit var binding: FragmentDemoContentABinding
     lateinit var viewModel: FragmentAViewModel
 
     override fun onCreateView(
@@ -34,7 +46,34 @@ class FragmentA : Fragment(), Injectable {
 
         Timber.d("Got injected fragment's own viewmodel instance: %s.", viewModel)
 
-        // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_demo_content, container, false)
+        // Inflate the layout for this fragment using data binding and set the view model
+        binding = DataBindingUtil.inflate(inflater, R.layout.fragment_demo_content_a, container, false)
+        binding.viewModel = viewModel
+        return binding.root
+    }
+
+    override fun onActivityCreated(savedInstanceState: Bundle?) {
+        super.onActivityCreated(savedInstanceState)
+        observeNavigationEvents(viewModel)
+    }
+
+    /**
+     * TODO: This is an example of how LiveData can be used to navigate. Update accordingly.
+     */
+    private fun observeNavigationEvents(viewModel: FragmentAViewModel) {
+        viewModel.featureXEvent.observe(this, Observer {
+            Timber.i("Launching feature X activity.")
+            startActivity(Intent(activity, FeatureXActivity::class.java))
+        })
+
+        viewModel.featureYEvent.observe(this, Observer {
+            Timber.i("Launching feature Y activity.")
+            startActivity(Intent(activity, FeatureYActivity::class.java))
+        })
+
+        viewModel.featureZEvent.observe(this, Observer {
+            Timber.i("Launching feature Z activity.")
+            startActivity(Intent(activity, FeatureZActivity::class.java))
+        })
     }
 }
